DEV: get competition on event details screen

+++ feature/events/src/main/java/com/rodionov/events/data/main/EventsAction.kt
@@ -1,5 +1,5 @@
 package com.rodionov.events.data.main
 
 sealed class EventsAction {
-    data class EventClick(val eventId: Long) : EventsAction()
+    data class EventClick(val eventId: Int) : EventsAction()
 }
